add comment function and add many excetion code and ajaxResquest and community.js

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -24,7 +24,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -46,7 +46,7 @@ public class PublishController {
             @RequestParam(value = "tag",required = false) String tag,
             HttpServletRequest request,
             Model model,
-            @RequestParam(value = "id",required = false)Integer id
+            @RequestParam(value = "id",required = false)Long id
                             ) {
         model.addAttribute("title",title);
         model.addAttribute("description",description);
